% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.R
\name{CommAttrWidget}
\alias{CommAttrWidget}
\title{Widget backed by a Jupyter Comm and a CRDT ydoc, with map-attr storage}
\description{
Opens a Jupyter Comm on the `"ywidget"` target and runs the Y.js sync
protocol on top of [YAttrWidget]'s map-attr storage. Constructor:
`CommAttrWidget$new(ydoc = NULL, comm_metadata = NULL)`.
}
\section{Super classes}{
\code{WidgetBase} -> \code{YAttrWidget} -> \code{.CommAttrWidgetBase} -> \code{CommAttrWidget}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-CommAttrWidget-clone}{\code{CommAttrWidget$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><code>WidgetBase$connect()</code></li>
  <li><code>WidgetBase$with_read()</code></li>
  <li><code>WidgetBase$with_write()</code></li>
  <li><code>YAttrWidget$register_storage()</code></li>
  <li><code>.CommAttrWidgetBase$comm_id()</code></li>
  <li><code>.CommAttrWidgetBase$initialize()</code></li>
  <li><code>.CommAttrWidgetBase$on_remote_message()</code></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommAttrWidget-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CommAttrWidget-clone}{}}}
\subsection{\code{CommAttrWidget$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CommAttrWidget$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
