% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.R
\name{CommRootWidget}
\alias{CommRootWidget}
\title{Widget backed by a Jupyter Comm and a CRDT ydoc, with root-CRDT storage}
\description{
Like [CommAttrWidget] but on top of [YRootWidget]'s per-root storage.
Constructor: `CommRootWidget$new(ydoc = NULL, comm_metadata = NULL)`.
}
\section{Super classes}{
\code{\link[ywidgets:WidgetBase]{WidgetBase}} -> \code{\link[ywidgets:YRootWidget]{YRootWidget}} -> \code{.CommRootWidgetBase} -> \code{CommRootWidget}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-CommRootWidget-clone}{\code{CommRootWidget$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="ywidgets" data-topic="WidgetBase" data-id="connect"><a href='../../ywidgets/html/WidgetBase.html#method-WidgetBase-connect'><code>WidgetBase$connect()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ywidgets" data-topic="YRootWidget" data-id="register_storage"><a href='../../ywidgets/html/YRootWidget.html#method-YRootWidget-register_storage'><code>YRootWidget$register_storage()</code></a></span></li>
  <li><code>.CommRootWidgetBase$comm_id()</code></li>
  <li><code>.CommRootWidgetBase$initialize()</code></li>
  <li><code>.CommRootWidgetBase$on_remote_message()</code></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommRootWidget-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CommRootWidget-clone}{}}}
\subsection{\code{CommRootWidget$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CommRootWidget$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
