% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.R
\name{CommWidget}
\alias{CommWidget}
\title{Widget backed by a Jupyter Comm and a CRDT ydoc}
\description{
Extends [Widget] by opening a Jupyter Comm and running the Y.js sync
protocol over it, mirroring `ypywidgets.CommWidget`: on `SyncStep1` reply
with a `SyncStep2` diff; on `SyncStep2` apply the peer's state and (once)
start forwarding local changes as `Update`s; on `Update` apply the peer's
incremental change.
}
\section{Super class}{
\code{\link[ywidget:Widget]{Widget}} -> \code{CommWidget}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-CommWidget-initialize}{\code{CommWidget$new()}}
    \item \href{#method-CommWidget-on_remote_message}{\code{CommWidget$on_remote_message()}}
    \item \href{#method-CommWidget-comm_id}{\code{CommWidget$comm_id()}}
    \item \href{#method-CommWidget-clone}{\code{CommWidget$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="ywidget" data-topic="Widget" data-id="model_name"><a href='../../ywidget/html/Widget.html#method-Widget-model_name'><code>Widget$model_name()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommWidget-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-CommWidget-initialize}{}}}
\subsection{\code{CommWidget$new()}}{
  Open a Comm on the `"ywidget"` target and send `SyncStep1`.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CommWidget$new(ydoc = NULL, comm_metadata = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{ydoc}}{Optional existing `yr::Doc` to adopt.}
      \item{\code{comm_metadata}}{Overrides the default metadata sent on comm open.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommWidget-on_remote_message"></a>}}
\if{latex}{\out{\hypertarget{method-CommWidget-on_remote_message}{}}}
\subsection{\code{CommWidget$on_remote_message()}}{
  Dispatch an incoming `yr::Message` to its sync handler.
  Public so hera callbacks can re-enter via a captured `self`, where R6
  bindings are otherwise broken.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CommWidget$on_remote_message(msg)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{msg}}{A `yr::Message`.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommWidget-comm_id"></a>}}
\if{latex}{\out{\hypertarget{method-CommWidget-comm_id}{}}}
\subsection{\code{CommWidget$comm_id()}}{
  The underlying comm id (used to build the display payload).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CommWidget$comm_id()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommWidget-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CommWidget-clone}{}}}
\subsection{\code{CommWidget$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CommWidget$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
