% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive.R
\name{LocalStorage}
\alias{LocalStorage}
\title{Default in-memory storage backend for `Reactive`.}
\description{
Default in-memory storage backend for `Reactive`.
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LocalStorage-initialize}{\code{LocalStorage$new()}}
    \item \href{#method-LocalStorage-read}{\code{LocalStorage$read()}}
    \item \href{#method-LocalStorage-update}{\code{LocalStorage$update()}}
    \item \href{#method-LocalStorage-clone}{\code{LocalStorage$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalStorage-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-LocalStorage-initialize}{}}}
\subsection{\code{LocalStorage$new()}}{
  Create the backend.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LocalStorage$new(value = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{value}}{Initial value.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalStorage-read"></a>}}
\if{latex}{\out{\hypertarget{method-LocalStorage-read}{}}}
\subsection{\code{LocalStorage$read()}}{
  Return the stored value.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LocalStorage$read()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalStorage-update"></a>}}
\if{latex}{\out{\hypertarget{method-LocalStorage-update}{}}}
\subsection{\code{LocalStorage$update()}}{
  Store `value`; no-op when unchanged.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LocalStorage$update(value)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{value}}{New value.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    `TRUE` iff the value was replaced.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalStorage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LocalStorage-clone}{}}}
\subsection{\code{LocalStorage$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LocalStorage$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
