% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive.R
\name{RemoteLocalStorage}
\alias{RemoteLocalStorage}
\title{Storage backend with a remote-change [Signal].}
\description{
Extends [LocalStorage] with a public `remote_changed` signal fired when the
value changes for reasons outside of `update()`. The in-memory backend has
no external source, so `remote_changed` never fires; subclasses backed by
shared state emit on it when they observe remote updates.
}
\section{Super class}{
\code{\link[ywidgets:LocalStorage]{LocalStorage}} -> \code{RemoteLocalStorage}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{remote_changed}}{[Signal] fired on remote changes to the value.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-RemoteLocalStorage-initialize}{\code{RemoteLocalStorage$new()}}
    \item \href{#method-RemoteLocalStorage-clone}{\code{RemoteLocalStorage$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="ywidgets" data-topic="LocalStorage" data-id="read"><a href='../../ywidgets/html/LocalStorage.html#method-LocalStorage-read'><code>LocalStorage$read()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ywidgets" data-topic="LocalStorage" data-id="update"><a href='../../ywidgets/html/LocalStorage.html#method-LocalStorage-update'><code>LocalStorage$update()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteLocalStorage-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteLocalStorage-initialize}{}}}
\subsection{\code{RemoteLocalStorage$new()}}{
  Create the backend.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RemoteLocalStorage$new(value = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{value}}{Initial value.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteLocalStorage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteLocalStorage-clone}{}}}
\subsection{\code{RemoteLocalStorage$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RemoteLocalStorage$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
