% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive.R
\name{Signal}
\alias{Signal}
\title{Event emitter.}
\description{
Broadcasts a value to all registered callbacks. Holds no state of its own.
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Signal-connect}{\code{Signal$connect()}}
    \item \href{#method-Signal-emit}{\code{Signal$emit()}}
    \item \href{#method-Signal-clone}{\code{Signal$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Signal-connect"></a>}}
\if{latex}{\out{\hypertarget{method-Signal-connect}{}}}
\subsection{\code{Signal$connect()}}{
  Register a callback invoked on every `emit()`.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Signal$connect(f)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{f}}{A function accepting the emitted arguments.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Signal-emit"></a>}}
\if{latex}{\out{\hypertarget{method-Signal-emit}{}}}
\subsection{\code{Signal$emit()}}{
  Invoke every registered callback with `...`.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Signal$emit(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Arguments forwarded to each callback.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Signal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Signal-clone}{}}}
\subsection{\code{Signal$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Signal$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
